Test collection move actions
