+++ client-v2/src/util/collectionUtils.js
@@ -1,6 +1,39 @@
 // @flow
 
 import { getURLCAPIID } from 'util/CAPIUtils';
+import {
+  removeGroupArticleFragment,
+  addGroupArticleFragment
+} from 'actions/Collections';
+import {
+  removeSupportingArticleFragment,
+  addSupportingArticleFragment
+} from 'actions/ArticleFragments';
+
+const getMoveActions = ({ payload: { id, from, to } }) => {
+  const getFromAction = () => {
+    if (from.parent.type === 'articleFragment') {
+      return removeSupportingArticleFragment(from.parent.id, id);
+    }
+
+    if (from.parent.type === 'group') {
+      return removeGroupArticleFragment(from.parent.id, id);
+    }
+    return () => null;
+  };
+
+  const getToAction = () => {
+    if (to.parent.type === 'articleFragment') {
+      return addSupportingArticleFragment(to.parent.id, id, to.index);
+    }
+    if (to.parent.type === 'group') {
+      return addGroupArticleFragment(to.parent.id, id, to.index);
+    }
+    return () => null;
+  };
+
+  return [getFromAction(), getToAction()];
+};
 
 const urlToArticle = (text: string) => {
   const id = getURLCAPIID(text);
@@ -13,4 +46,4 @@ const urlToArticle = (text: string) => {
     : 'Can`t covert text to article';
 };
 
-export { urlToArticle };
+export { urlToArticle, getMoveActions };
